% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thlcol.R
\name{thlcol}
\alias{thlcol}
\alias{thlcolbynum}
\alias{thlgrey}
\title{Vector of THL colors}
\usage{
thlcol(shadegroup = NULL, darkness = NULL, alpha = 1)

thlgrey(darkness = NULL, alpha = 1)

thlcolbynum(i = 1, alpha = 1)
}
\arguments{
\item{shadegroup}{group of groups of shades to use, a letter from "A" to "I"}

\item{darkness}{darkness or darknesses to use, "V" light, "K" medium, "T" dark, "L" bright (use "L" sparingly) for thlcol, 1 to 5 for thlgrey}

\item{alpha}{opacity of the color (recycled)}

\item{i}{number of the color (1-41)}
}
\value{
a vector of RGB colors
}
\description{
Create a vector of THL colors as defined in https://yhteistyotilat.fi/wiki08/x/DgBJAQ
The parameters are recycled
}
\section{Functions}{
\itemize{
\item \code{thlgrey}: Vector of THL greys

\item \code{thlcolbynum}: Vector of THL colors by number
}}

