% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thlcol.R
\name{thlcolsample}
\alias{thlcolsample}
\alias{thlcolsamplebar}
\title{Demo function}
\usage{
thlcolsample()

thlcolsamplebar(...)
}
\arguments{
\item{...}{parameters passed to thlpal}
}
\description{
Demonstrate the use of THL colors
}
\section{Functions}{
\itemize{
\item \code{thlcolsamplebar}: Draw single bar with a palette
}}

