% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thlcol.R
\name{thlpal}
\alias{thlpal}
\title{Create THL colorpalette}
\usage{
thlpal(palette = c("grey1", "grey2", "shadegroup", "large"),
  shadegroup = "A", altshadegroup = NULL, darkness = c("K", "V", "T",
  "L"), n = NULL, m = NULL, alpha = 1)
}
\arguments{
\item{palette}{type of the palette}

\item{shadegroup}{base shade group, from "A" to "I"}

\item{altshadegroup}{base shade group, from "A" to "I" (used by palette "grey2")}

\item{darkness}{base darkness}

\item{n}{number of colors for palette "large"}

\item{m}{maximum number of colors. If m>n, pad with NA}

\item{alpha}{opacity (recycled)}
}
\value{
a vector of RGB colors
}
\description{
Create a color palette using the rules in https://yhteistyotilat.fi/wiki08/x/DgBJAQ
}

